+++ dotnet/src/Winium.Elements.Desktop/DesktopElement.cs
@@ -6,18 +6,17 @@
     using System.Collections.Generic;
 
     using OpenQA.Selenium;
-    using OpenQA.Selenium.Remote;
-
+    using OpenQA.Selenium.Winium;
     using Winium.Elements.Desktop.Extensions;
 
     #endregion
 
-    public abstract class DesktopElement : RemoteWebElement
+    public abstract class DesktopElement : WiniumElement
     {
         #region Constructors and Destructors
 
-        protected DesktopElement(IWebElement element)
-            : base(GetRemoteWebDriver(element), element.GetId())
+        protected DesktopElement(WiniumElement element)
+            : base(GetWiniumDriver(element), element.GetId())
         {
         }
 
@@ -25,7 +24,7 @@
 
         #region Methods
 
-        protected RemoteWebElement CreateRemoteWebElementFromResponse(Response response)
+        protected WiniumElement CreateWiniumElementFromResponse(Response response)
         {
             var elementDictionary = response.Value as Dictionary<string, object>;
             if (elementDictionary == null)
@@ -33,18 +32,18 @@
                 return null;
             }
 
-            return new RemoteWebElement((RemoteWebDriver)this.WrappedDriver, (string)elementDictionary["ELEMENT"]);
+            return new WiniumElement((WiniumDriver)this.WrappedDriver, (string)elementDictionary["element-6066-11e4-a52e-4f735466cecf"]);
         }
 
-        private static RemoteWebDriver GetRemoteWebDriver(IWebElement element)
+        private static WiniumDriver GetWiniumDriver(WiniumElement element)
         {
-            var remoteWebElement = element as RemoteWebElement;
-            if (remoteWebElement == null)
+            var winiumElement = element as WiniumElement;
+            if (winiumElement == null)
             {
-                throw new InvalidCastException("Specified cast is not valid. Please use RemoteWebElement as parameter.");
+                throw new InvalidCastException("Specified cast is not valid. Please use WiniumElement as parameter.");
             }
 
-            return (RemoteWebDriver)remoteWebElement.WrappedDriver;
+            return (WiniumDriver)winiumElement.WrappedDriver;
         }
 
         #endregion