+++ dotnet/CHANGELOG.md
@@ -2,10 +2,15 @@
 
 <!--## Unreleased-->
 
+## v1.0.0
+- Support Selenium.WebDriver bindings version 4.11.0
+- Support Winium-WebDriver bindings version 1.0.0
+
 ## v0.3.0
 - Support Selenium.WebDriver bindings version 3.141.0
 - Use Winium-WebDriver nuget package (create WiniumDriver driver in your tests)
-## v0.2.0
+
+## v0.2.0
 
 - Support Selenium.WebDriver bindings version 2.8.0
 - Use Winium.WebDriver nuget package (create WiniumDriver driver in your tests)
@@ -19,5 +24,6 @@
 - menu (find item, select item)
 
 
-
+
+